[ALPS04966423] irq: add hotplug callback for prefer irq
[Detail]
add hotplug callback for prefer irq

MTK-Commit-Id: 4be29b726763a51a6b44029e72a9d1415b20248d

CR-Id: ALPS04966423
Feature: System Performance

+++ include/linux/irqchip/mtk-gic-extend.h
@@ -85,8 +85,8 @@ void irq_raise_softirq(const struct cpumask *mask, unsigned int irq);
 void gic_set_primask(void);
 /* restore the priority mask value */
 void gic_clear_primask(void);
-int add_cpu_to_prefer_schedule_domain(unsigned long cpu);
-int remove_cpu_from_prefer_schedule_domain(unsigned long cpu);
+int add_cpu_to_prefer_schedule_domain(unsigned int cpu);
+int remove_cpu_from_prefer_schedule_domain(unsigned int cpu);
 
 #ifdef CONFIG_MTK_SYSIRQ
 static inline struct irq_data *get_gic_irq_data(struct irq_data *d)
